pyproject.toml
[project]
name = "ai-proxy"
version = "0.1.0"
description = ""
authors = [
  { name = "Marwa Alouani", email = "malouani@naoplay.fr" }
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = ["annotated-doc (==0.0.4)", "annotated-types (==0.7.0)", "anyio (==4.12.1)", "fastapi (==0.133.0)", "filelock (==3.24.3)", "fsspec (==2026.2.0)", "idna (==3.11)", "jinja2 (==3.1.6)", "markupsafe (==3.0.3)", "mpmath (==1.3.0)", "pydantic (==2.12.5)", "pydantic-core (==2.41.5)", "python-dateutil (==2.9.0.post0)", "setuptools (==82.0.0)", "six (==1.17.0)", "starlette (==0.52.1)", "sympy (==1.14.0)", "typing-inspection (==0.4.2)", "typing-extensions (==4.15.0)", "litellm (==1.81.15)", "uvicorn (==0.41.0)", "openai (==2.24.0)", "httpx (==0.28.1)", "requests (==2.32.5)", "python-dotenv (==1.2.1)", "pyyaml (==6.0.3)", "rich (==14.3.3)", "typer (==0.24.1)", "fastuuid (==0.14.0)", "google-cloud-secret-manager (>=2.26.0,<3.0.0)", "google-crc32c (>=1.8.0,<2.0.0)"]

[tool.poetry]
packages = [{ include = "proxy" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [".git", ".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"